+++ spinetoolbox/spine_db_editor/widgets/tree_view_mixin.py
@@ -11,7 +11,8 @@
 ######################################################################################################################
 
 """Contains the TreeViewMixin class."""
-from PySide6.QtCore import Slot
+from PySide6.QtCore import Slot, Qt, QEvent
+from PySide6.QtGui import QMouseEvent
 from .add_items_dialogs import (
     AddEntityClassesDialog,
     AddEntitiesDialog,
@@ -33,7 +34,7 @@ from ...spine_db_parcel import SpineDBParcel
 
 
 class TreeViewMixin:
-    """Provides object and relationship trees for the Spine db editor."""
+    """Provides entity, alternative, scenario and parameter value list trees for the Spine db editor."""
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
@@ -52,6 +53,134 @@ class TreeViewMixin:
             view.setModel(model)
             view.connect_spine_db_editor(self)
             view.header().setResizeContentsPrecision(self.visible_rows)
+        self.clear_tree_selections = True
+        # Filter caches
+        self._filter_class_ids = {}  # Class ids from entity class- and entity selections (cascading)
+        self._filter_entity_ids = {}  # Entity ids from entity selections
+        self._filter_alternative_ids = {}  # Alternative ids
+        self._filter_scenario_ids = {}  # Scenario ids by db_map. Each scenario id maps to alternatives sorted by rank.
+        self._filter_parameter_value_ids = {}  # Entity ids for currently accepted parameter value rows in tables
+
+    def connect_signals(self):
+        """Connects the signals"""
+        super().connect_signals()
+        self.ui.treeView_entity.tree_selection_changed.connect(self._handle_entity_tree_selection_changed)
+        self.ui.alternative_tree_view.alternative_selection_changed.connect(self._handle_alternative_selection_changed)
+        self.ui.scenario_tree_view.scenario_selection_changed.connect(
+            self._handle_scenario_alternative_selection_changed
+        )
+        self.entity_alternative_model.dataChanged.connect(self.build_graph)
+        self.parameter_value_model.dataChanged.connect(self._refresh_parameters_and_graph)
+
+    def _refresh_parameters_and_graph(self):
+        self._update_filter_parameter_value_ids()
+        self.build_graph()
+
+    @Slot(dict)
+    def _handle_entity_tree_selection_changed(self, selected_indexes):
+        entity_indexes = set(selected_indexes.get("entity", {}).keys())
+        entity_indexes |= {
+            parent
+            for parent in (i.parent() for i in entity_indexes)
+            if self.entity_tree_model.item_from_index(parent).item_type == "entity"
+        }
+        entity_class_indexes = set(selected_indexes.get("entity_class", {}).keys()) | {
+            parent_ind
+            for parent_ind in (ind.parent() for ind in entity_indexes)
+            if self.entity_tree_model.item_from_index(parent_ind).item_type == "entity_class"
+        }
+        self._filter_class_ids = self._db_map_ids(entity_class_indexes)
+        self._filter_entity_ids = self.db_mngr.db_map_class_ids(group_items_by_db_map(entity_indexes))
+        # View specific stuff:
+        self._update_selected_item_type_db_map_ids(selected_indexes)
+        self._reset_filters()
+        self._set_default_parameter_data(self.ui.treeView_entity.selectionModel().currentIndex())
+        self._update_filter_parameter_value_ids()
+        self.build_graph()
+
+    @Slot(dict)
+    def _handle_alternative_selection_changed(self, selected):
+        self._filter_alternative_ids.clear()
+        for db_map, alt_ids in selected.items():
+            if not alt_ids:
+                continue
+            self._filter_alternative_ids.setdefault(db_map, set()).update(alt_ids)
+        # View specific stuff:
+        self._reset_filters()
+        self._update_filter_parameter_value_ids()
+        self.build_graph()
+
+    @Slot(dict)
+    def _handle_scenario_alternative_selection_changed(self, selected_ids):
+        self._filter_scenario_ids = selected_ids
+        # View specific stuff:
+        self._reset_filters()
+        self._update_filter_parameter_value_ids()
+        self.build_graph()
+
+    def _update_filter_parameter_value_ids(self):
+        """Updates the parameter"""
+        for db_map in self.db_maps:
+            single_models = self.parameter_value_model._models_with_db_map(db_map)
+            if not single_models:
+                return True
+            parameter_value_ids = set()
+            for model in single_models:
+                for _, row in self.parameter_value_model._row_map_iterator_for_model(model):
+                    parameter_value_ids.add(model._db_item(row).get("entity_id"))
+            if not parameter_value_ids and db_map in self._filter_parameter_value_ids:
+                del self._filter_parameter_value_ids[db_map]
+            self._filter_parameter_value_ids[db_map] = parameter_value_ids
+
+    def handle_mousepress(self, tree_view, event):
+        """Overrides selection behaviour if the user has selected sticky selection in Settings.
+        If sticky selection is enabled, multiple-selection is enabled when selecting items in the Object tree.
+        Pressing the Ctrl-button down, enables single selection.
+
+        Args:
+            tree_view (QTreeView): The treeview where the mouse click was in.
+            event (QMouseEvent): event
+        """
+        if tree_view == self.ui.treeView_parameter_value_list:
+            return event
+        self.clear_tree_selections = True
+        sticky_selection = self.qsettings.value("appSettings/stickySelection", defaultValue="false")
+        if sticky_selection == "false":
+            pos = tree_view.viewport().mapFromGlobal(event.globalPos())
+            index = tree_view.indexAt(pos)
+            modifiers = event.modifiers()
+            if modifiers & Qt.ControlModifier:
+                self.clear_tree_selections = False
+            elif not (tree_view.selectionModel().hasSelection() or index.isValid()):
+                # Ensure selection clearing when empty space is clicked on a tree that doesn't have selections.
+                self._clear_all_other_selections(tree_view)
+            return event
+        local_pos = event.position()
+        window_pos = event.scenePosition()
+        screen_pos = event.globalPosition()
+        button = event.button()
+        buttons = event.buttons()
+        modifiers = event.modifiers()
+        if modifiers & Qt.ControlModifier:
+            modifiers &= ~Qt.ControlModifier
+        else:
+            modifiers |= Qt.ControlModifier
+            self.clear_tree_selections = False
+        source = event.source()
+        new_event = QMouseEvent(
+            QEvent.MouseButtonPress, local_pos, window_pos, screen_pos, button, buttons, modifiers, source
+        )
+        return new_event
+
+    def _clear_all_other_selections(self, current):
+        """Clears all selections from other tree views except from the current one.
+
+        Args:
+            current: the tree where the selection that was made
+        """
+        for tree in [self.ui.treeView_entity, self.ui.scenario_tree_view, self.ui.alternative_tree_view]:
+            if tree != current:
+                tree.selectionModel().clearSelection()
 
     def init_models(self):
         """Initializes models."""